# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Penjualan",
        "description": "API untuk manajemen data penjualan",
        "contact": {
            "name": "Muhammad Gibran",
            "email": "gibran.devv@email.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pelanggan/{id}": {
            "get": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Mendapatkan detail pelanggan berdasarkan ID",
                "operationId": "1a8a90bb364c57f82cc2c44b1c385e75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan detail pelanggan"
                    },
                    "404": {
                        "description": "Pelanggan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Memperbarui data pelanggan",
                "operationId": "bdae5065b6bd1643790d30366c079c22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data pelanggan yang diperbarui",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_pelanggan": {
                                        "type": "string",
                                        "example": "John Doe Updated"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Contoh Diperbarui No.456"
                                    },
                                    "telepon": {
                                        "type": "string",
                                        "example": "087654321098"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil memperbarui data pelanggan"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Menghapus pelanggan",
                "operationId": "37d2e71208bceecb7cf5417e6305a451",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pelanggan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menghapus pelanggan"
                    },
                    "404": {
                        "description": "Pelanggan tidak ditemukan"
                    }
                }
            }
        },
        "/api/pelanggan/create": {
            "post": {
                "tags": [
                    "Pelanggan"
                ],
                "summary": "Membuat pelanggan baru",
                "operationId": "498fce8e34caf34ad8c9490f69a330f9",
                "requestBody": {
                    "description": "Data pelanggan baru",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_pelanggan": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Contoh No.123"
                                    },
                                    "telepon": {
                                        "type": "string",
                                        "example": "081234567890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Berhasil membuat pelanggan baru"
                    },
                    "400": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
    },
...
}